Move IterableEndpoint fixture and test into their own package
The endpoint generator testing setup requires a separate package for
each test to avoid interference between tests. This is because the
generator in the tests emits all classes per package.
